pyproject.toml
[project]
name = "open-conference-url"
version = "7.3.0"
description = "An Alfred workflow to quickly open conference links from upcoming calendar events."
readme = "README.md"
requires-python = ">=3.9"
authors = [
    {name = "Caleb Evans", email = "caleb@calebevans.me"}
]
maintainers = [
    {name = "Caleb Evans", email = "caleb@calebevans.me"}
]
license = "MIT"
dependencies = []

[dependency-groups]
dev = [
    "alfred-workflow-packager==3.1.1",
    "freezegun==1.5.2",
    "mypy==1.4.1",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "python-dotenv==1.0.0",
    "ruff>=0.12.0",
]

[tool.ruff]
include = ["**/*.py"]

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    "W",
    # pyflakes
    "F",
    # isort
    "I",
    # perflint
    "PERF",
]

[tool.coverage.run]
branch = true
# Only check coverage for source files
include = ["ocu/**/*.py"]

[tool.coverage.report]
# Regexes for lines to exclude from consideration
exclude_lines = [
    "pragma: no cover",
    # Ignore non-runnable code
    "if __name__ == .__main__.:",
    "pass",
]

[tool.mypy]
disallow_untyped_defs = true

# It's not practical to add type annotations to tests
[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
